caught a few more failed merges

+++ src/fsort.c
@@ -175,8 +175,8 @@ SEXP fsort(SEXP x, SEXP verboseArg) {
   // TODO: change all calloc, malloc and free to Calloc and Free to be robust to error() and catch ooms.
 
   if (verbose) Rprintf(_("counts is %dMB (%d pages per nBatch=%d, batchSize=%"PRIu64", lastBatchSize=%"PRIu64")\n"),
-                       nBatch*MSBsize*sizeof(R_xlen_t)/(1024*1024),
-                       nBatch*MSBsize*sizeof(R_xlen_t)/(4*1024*nBatch),
+                       (int)(nBatch*MSBsize*sizeof(R_xlen_t)/(1024*1024)),
+                       (int)(nBatch*MSBsize*sizeof(R_xlen_t)/(4*1024*nBatch)),
                        nBatch, (uint64_t)batchSize, (uint64_t)lastBatchSize);
   t[3] = wallclock();
   #pragma omp parallel for num_threads(nth)
